moved the last 2 loaders out of loaders into their own files
